refactor: extract user list logic to a separate component to be used in followers/following ...etc
